ADEN-8857 moving btl wrapper
