+++ app/modules/ads/index.js
@@ -6,7 +6,7 @@ import { fanTakeoverResolver } from './fan-takeover-resolver';
 import { adblockDetector } from './tracking/adblock-detector';
 import { pageTracker } from './tracking/page-tracker';
 import { biddersDelayer } from './bidders-delayer';
-import { billTheLizardWrapper } from './bill-the-lizard-wrapper';
+import { billTheLizardWrapper } from './ml/bill-the-lizard-wrapper';
 import { appEvents } from './events';
 import { logError } from '../event-logger';
 import { trackScrollY, trackXClick } from '../../utils/track';
